Moved bounceAudio to AudioFileUtils
